Added helpers and website into folder, restructure
